openness_score/coreboot.py: Handle FMAP layout problems
It is possible for the FMAP layout to be non-contiguous. Catch such
cases and count the non-described space as closed source.

Add SI_BIOS to the list of container regions, to avoid warnings.

Catch regions that may have numbers in the name.

Do not print a warning when a coreboot config is not found in a CBFS
region not expected to have a config file.

Signed-off-by: Michał Żygowski <michal.zygowski@3mdeb.com>