openness_score/coreboot.py: Handle ifdtool corner cases
Add boolean variable holding the IFD presence in the image.
Use it to fall back to regular parsing without ifdtool if
the variabel is is set to False.

Handle cases where the FMAP layout does not start at the
beginning of flash and there is no IFD in the image.

Signed-off-by: Michał Żygowski <michal.zygowski@3mdeb.com>